perbaikan import barang dari excel

+++ application/controllers/admin/Barang.php
@@ -1,4 +1,5 @@
 <?php
+
 class Barang extends CI_Controller{
 
 	function __construct(){
@@ -19,13 +20,16 @@ class Barang extends CI_Controller{
 	function index(){
 
 		if($this->session->userdata('akses')=='1'){
+
 			$data['data']=$this->m_barang->tampil_barang();
-			// $data['kat']=$this->m_kategori->tampil_kategori();
 			$data['kat2']=$this->m_kategori->tampil_kategori();
 			$data['uom']=$this->m_uom->tampil_uom();
 			$this->load->view('admin/v_barang',$data);
+		
 		}else{
+
 	        echo "Halaman tidak ditemukan";
+	    
 	    }
 
 	}
@@ -33,6 +37,7 @@ class Barang extends CI_Controller{
 	function tambah_barang(){
 
 		if($this->session->userdata('akses')=='1'){
+			
 			// $kobar=$this->m_barang->get_kobar();
 			$kobar=$this->input->post('kobar');
 			$nabar=$this->input->post('nabar');
@@ -43,16 +48,32 @@ class Barang extends CI_Controller{
 			$harjul_grosir=str_replace(',', '', $this->input->post('harjul_grosir'));
 			$stok=$this->input->post('stok');
 			$min_stok=$this->input->post('min_stok');
-			$simpan = $this->m_barang->simpan_barang($kobar, $nabar, $kat, $satuan, $harpok, $harjul, $harjul_grosir, $stok, $min_stok);
 
-			// audit 
-			$this->m_global->audit_master('Insert', 'tbl_barang', $simpan);
-			// audit 
+			// cek kode
+			if($this->m_barang->cek_kode_exist($kobar) == TRUE){
 
-			echo $this->session->set_flashdata('msg','<label class="label label-success">Barang '.$nabar.' Berhasil Ditambahkan </label>');
-			redirect('admin/barang');
+				$simpan = $this->m_barang->simpan_barang($kobar, $nabar, $kat, $satuan, $harpok, $harjul, $harjul_grosir, $stok, $min_stok);
+
+				// audit 
+				$this->m_global->audit_master('Insert', 'tbl_barang', $simpan);
+				// audit 
+
+				echo $this->session->set_flashdata('msg','<label class="label label-success">Barang '.$nabar.' Berhasil Ditambahkan </label>');
+				redirect('admin/barang');
+			
+			}else{
+
+				echo $this->session->set_flashdata('msg','<label class="label label-danger">Kode Barang '.$kobar.' Sudah Ada ! </label>');
+				redirect('admin/barang');
+
+			}
+
+			// cek kode
+		
 		}else{
+
 	        echo "Halaman tidak ditemukan";
+	    
 	    }
 
 	}
